# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus AI Platform - AI Recruitment & CFO SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job postings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /payroll/runs:
    get:
      operationId: listPayrollRuns
      tags: [payroll]
      summary: List payroll runs
      responses:
        "200":
          description: List of payroll runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRun"
    post:
      operationId: createPayrollRun
      tags: [payroll]
      summary: Create a new payroll run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunInput"
      responses:
        "201":
          description: Created payroll run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}:
    get:
      operationId: getPayrollRun
      tags: [payroll]
      summary: Get payroll run by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payroll run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
    patch:
      operationId: updatePayrollRun
      tags: [payroll]
      summary: Update payroll run status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}/process:
    post:
      operationId: processPayrollRun
      tags: [payroll]
      summary: Process and finalize a payroll run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Processed payroll run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, department, location, employment_type, status, salary_min, salary_max, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employment_type:
          type: string
          enum: [full_time, part_time, contract, internship]
        status:
          type: string
          enum: [open, closed, on_hold, draft]
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        salary_min:
          type: number
        salary_max:
          type: number
        candidate_count:
          type: integer
        posted_by:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]
    JobInput:
      type: object
      required: [title, department, location, employment_type, salary_min, salary_max]
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employment_type:
          type: string
          enum: [full_time, part_time, contract, internship]
        status:
          type: string
          enum: [open, closed, on_hold, draft]
        description:
          type: string
        requirements:
          type: string
        salary_min:
          type: number
        salary_max:
          type: number
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employment_type:
          type: string
        status:
          type: string
        description:
          type: string
        requirements:
          type: string
        salary_min:
          type: number
        salary_max:
          type: number
...
    PayrollRun:
      type: object
      required: [id, period_start, period_end, status, total_gross, total_net, total_tax, employee_count, created_at]
      properties:
        id:
          type: string
        period_start:
          type: string
        period_end:
          type: string
        status:
          type: string
          enum: [draft, processing, completed, approved]
        total_gross:
          type: number
        total_net:
          type: number
        total_tax:
          type: number
        total_benefits:
          type: number
        employee_count:
          type: integer
        processed_at:
          type: ["string", "null"]
        approved_by:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
    PayrollRunInput:
      type: object
      required: [period_start, period_end]
      properties:
        period_start:
          type: string
        period_end:
          type: string
        notes:
          type: string
    PayrollRunUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
...